Handle wifi password on failure
The logic of password clearing and user requests
is currently happening after any retry under the
maximum retries. This patch move it to the path
of failure, to guarantee that only on multiple
retries the password is requested to the user.

The patch also removes a prompt when a single
association fail.

Signed-off-by: José Pekkarinen <jose.pekkarinen@foxhound.fi>